refactor: set staticNetwork in erc20 balance fetcher - split cosmos/fetch
